requirements.txt
asgiref==3.11.1
Django
djangorestframework==3.16.1
sqlparse==0.5.5
typing_extensions==4.15.0
django-extensions
jsonschema